refactoring comms packages
