+++ kernel/packages/shared/comms/v1/BrokerConnection.ts
@@ -13,10 +13,10 @@ import { Auth } from 'shared/auth/Auth'
 import { AuthData } from './proto/comms'
 import { Message } from 'google-protobuf'
 import { commConfigurations } from 'config'
-import { Stats } from '../comms/debug'
+import { Stats } from '../../comms/debug'
 import { IBrokerConnection, BrokerMessage, SocketReadyState } from './IBrokerConnection'
 import { ILogger, createLogger } from 'shared/logger'
-import { Observable } from '../../decentraland-ecs/src/ecs/Observable'
+import { Observable } from 'decentraland-ecs/src/ecs/Observable'
 
 export class BrokerConnection implements IBrokerConnection {
   public alias: string | null = null
