+++ kernel/packages/shared/comms/v1/IBrokerConnection.ts
@@ -1,5 +1,5 @@
-import { Observable } from '../../decentraland-ecs/src/ecs/Observable'
-import { Stats } from '../comms/debug'
+import { Observable } from 'decentraland-ecs/src/ecs/Observable'
+import { Stats } from '../../comms/debug'
 
 export type BrokerMessage = {
   data: Uint8Array
