+++ kernel/packages/ui/avatar/avatarSystem.ts
@@ -11,7 +11,7 @@ import {
   UserMessage,
   UserRemovedMessage,
   UUID
-} from 'shared/comms-interface/types'
+} from 'shared/comms/interface/types'
 import { execute } from './rpc'
 import { ProfileForRenderer } from 'decentraland-ecs/src/decentraland/Types'
